Moved RPCs from ExtendDeadlineScript to LGUStore as the gameObject was not being initialized
Will have to look more into the prefab if I'm missing some setting on it to not let it run Start() method
